refactor: 패키지 구조 수정, 주석, TODO 작성, jar생성
