--- src/main/java/org/prgrms/vouchermanager/voucher/MemoryVoucherRepository.java
@@ -1,29 +1,0 @@
-package org.prgrms.vouchermanager.voucher;
-
-import org.springframework.stereotype.Repository;
-
-import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
-
-@Repository
-public class MemoryVoucherRepository implements VoucherRepository{
-
-    private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
-    // TODO: 아주 만약 UUID가 겹친다면 예외처리?
-
-    @Override
-    public Optional<Voucher> findById(UUID voucherId) {
-        return Optional.ofNullable(storage.get(voucherId));
-    }
-
-    @Override
-    public Voucher insert(Voucher voucher) {
-        storage.put(voucher.getVoucherId(), voucher);
-        return voucher;
-    }
-
-    @Override
-    public List<Voucher> getAll() {
-        return new ArrayList<>(storage.values());
-    }
-}
