--- src/main/java/org/prgrms/vouchermanager/voucher/Voucher.java
@@ -1,9 +1,0 @@
-package org.prgrms.vouchermanager.voucher;
-
-import java.util.UUID;
-
-public interface Voucher {
-    UUID getVoucherId();
-    VoucherType getVoucherType();
-    long discount(long beforeDiscount);
-}
