--- src/main/java/org/prgrms/vouchermanager/voucher/VoucherRepository.java
@@ -1,11 +1,0 @@
-package org.prgrms.vouchermanager.voucher;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface VoucherRepository {
-    Optional<Voucher> findById(UUID voucherId);
-    Voucher insert(Voucher voucher);
-    List<Voucher> getAll();
-}
