+++ src/main/java/org/prgrms/vouchermanager/voucher/domain/Voucher.java
@@ -1,0 +1,9 @@
+package org.prgrms.vouchermanager.voucher.domain;
+
+import java.util.UUID;
+
+public interface Voucher {
+    UUID getVoucherId();
+    VoucherType getVoucherType();
+    long discount(long beforeDiscount);
+}
