+++ src/main/java/org/prgrms/vouchermanager/voucher/domain/VoucherFactory.java
@@ -1,0 +1,16 @@
+package org.prgrms.vouchermanager.voucher.domain;
+
+/**
+ * VoucherType에 따라 voucher객체를 반환하는 팩토리
+ */
+public class VoucherFactory {
+
+    public static Voucher getVoucher(VoucherType type, long amount) {
+        switch (type){
+            case FIXED -> {return new FixedAmountVoucher(amount);}
+            case PERCENT -> {return new PercentDiscountVoucher(amount);}
+        }
+        return new FixedAmountVoucher(amount);
+    }
+
+}
