+++ src/main/java/org/prgrms/vouchermanager/voucher/domain/VoucherType.java
@@ -1,0 +1,16 @@
+package org.prgrms.vouchermanager.voucher.domain;
+
+import java.util.stream.Stream;
+
+public enum VoucherType {
+    INVALID,
+    FIXED,
+    PERCENT;
+
+    public static VoucherType findVoucherType(String type) {
+        return Stream.of(values())
+                .filter(v -> type.equalsIgnoreCase(v.toString()))
+                .findFirst().orElse(INVALID);
+    }
+
+}
