+++ src/main/java/org/prgrms/vouchermanager/voucher/repository/MemoryVoucherRepository.java
@@ -1,0 +1,30 @@
+package org.prgrms.vouchermanager.voucher.repository;
+
+import org.prgrms.vouchermanager.voucher.domain.Voucher;
+import org.springframework.stereotype.Repository;
+
+import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Repository
+public class MemoryVoucherRepository implements VoucherRepository{
+
+    private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
+    // TODO: 아주 만약 UUID가 겹친다면 예외처리?
+
+    @Override
+    public Optional<Voucher> findById(UUID voucherId) {
+        return Optional.ofNullable(storage.get(voucherId));
+    }
+
+    @Override
+    public Voucher insert(Voucher voucher) {
+        storage.put(voucher.getVoucherId(), voucher);
+        return voucher;
+    }
+
+    @Override
+    public List<Voucher> getAll() {
+        return new ArrayList<>(storage.values());
+    }
+}
