+++ src/main/java/org/prgrms/vouchermanager/voucher/service/VoucherService.java
@@ -1,0 +1,16 @@
+package org.prgrms.vouchermanager.voucher.service;
+
+import org.prgrms.vouchermanager.voucher.domain.Voucher;
+import org.prgrms.vouchermanager.voucher.domain.VoucherType;
+
+import java.util.UUID;
+
+/**
+ * Voucher에 관한 서비스
+ */
+public interface VoucherService {
+    void createVoucher(VoucherType type, Long amount);
+    String allVouchersToString();
+    Voucher findVoucher(UUID voucherId);
+    void useVoucher(Voucher voucher);
+}
